requirements.txt
async-timeout==5.0.1
certifi==2025.4.26
charset-normalizer==3.4.2
fire==0.7.0
idna==3.10
mysql-connector-python==9.2.0
python-dotenv==1.1.0
redis==6.2.0
requests==2.32.3
ruff==0.11.11
termcolor==3.1.0
urllib3==2.4.0
